rename contact-creation-manager
