--- packages/twenty-server/src/modules/auto-companies-and-contacts-creation/jobs/auto-companies-and-contacts-creation-job.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-
-import { AutoCompaniesAndContactsCreationModule } from 'src/modules/auto-companies-and-contacts-creation/auto-companies-and-contacts-creation.module';
-import { CreateCompanyAndContactJob } from 'src/modules/auto-companies-and-contacts-creation/jobs/create-company-and-contact.job';
-
-@Module({
-  imports: [AutoCompaniesAndContactsCreationModule],
-  providers: [CreateCompanyAndContactJob],
-})
-export class AutoCompaniesAndContactsCreationJobModule {}
