+++ packages/twenty-server/src/modules/contact-creation-manager/constants/contacts-creation-batch-size.constant.ts
