+++ packages/twenty-server/src/modules/contact-creation-manager/create-contact/create-contact.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
 import { ObjectMetadataRepositoryModule } from 'src/engine/object-metadata-repository/object-metadata-repository.module';
-import { CreateContactService } from 'src/modules/auto-companies-and-contacts-creation/create-contact/create-contact.service';
+import { CreateContactService } from 'src/modules/contact-creation-manager/create-contact/create-contact.service';
 import { PersonWorkspaceEntity } from 'src/modules/person/standard-objects/person.workspace-entity';
 
 @Module({
