+++ packages/twenty-server/src/modules/contact-creation-manager/services/create-company-and-contact.service.ts
@@ -10,14 +10,14 @@ import { ObjectRecordCreateEvent } from 'src/engine/integrations/event-emitter/t
 import { ObjectMetadataEntity } from 'src/engine/metadata-modules/object-metadata/object-metadata.entity';
 import { InjectObjectMetadataRepository } from 'src/engine/object-metadata-repository/object-metadata-repository.decorator';
 import { STANDARD_OBJECT_IDS } from 'src/engine/workspace-manager/workspace-sync-metadata/constants/standard-object-ids';
-import { CONTACTS_CREATION_BATCH_SIZE } from 'src/modules/auto-companies-and-contacts-creation/constants/contacts-creation-batch-size.constant';
-import { CreateCompanyService } from 'src/modules/auto-companies-and-contacts-creation/create-company/create-company.service';
-import { CreateContactService } from 'src/modules/auto-companies-and-contacts-creation/create-contact/create-contact.service';
-import { Contact } from 'src/modules/auto-companies-and-contacts-creation/types/contact.type';
-import { filterOutSelfAndContactsFromCompanyOrWorkspace } from 'src/modules/auto-companies-and-contacts-creation/utils/filter-out-contacts-from-company-or-workspace.util';
-import { getDomainNameFromHandle } from 'src/modules/auto-companies-and-contacts-creation/utils/get-domain-name-from-handle.util';
-import { getUniqueContactsAndHandles } from 'src/modules/auto-companies-and-contacts-creation/utils/get-unique-contacts-and-handles.util';
 import { ConnectedAccountWorkspaceEntity } from 'src/modules/connected-account/standard-objects/connected-account.workspace-entity';
+import { CONTACTS_CREATION_BATCH_SIZE } from 'src/modules/contact-creation-manager/constants/contacts-creation-batch-size.constant';
+import { CreateCompanyService } from 'src/modules/contact-creation-manager/create-company/create-company.service';
+import { CreateContactService } from 'src/modules/contact-creation-manager/create-contact/create-contact.service';
+import { Contact } from 'src/modules/contact-creation-manager/types/contact.type';
+import { filterOutSelfAndContactsFromCompanyOrWorkspace } from 'src/modules/contact-creation-manager/utils/filter-out-contacts-from-company-or-workspace.util';
+import { getDomainNameFromHandle } from 'src/modules/contact-creation-manager/utils/get-domain-name-from-handle.util';
+import { getUniqueContactsAndHandles } from 'src/modules/contact-creation-manager/utils/get-unique-contacts-and-handles.util';
 import { PersonRepository } from 'src/modules/person/repositories/person.repository';
 import { PersonWorkspaceEntity } from 'src/modules/person/standard-objects/person.workspace-entity';
 import { WorkspaceMemberRepository } from 'src/modules/workspace-member/repositories/workspace-member.repository';
