+++ packages/twenty-server/src/modules/contact-creation-manager/utils/__tests__/get-unique-contacts-and-handles.spec.ts
@@ -1,5 +1,5 @@
-import { Contact } from 'src/modules/auto-companies-and-contacts-creation/types/contact.type';
-import { getUniqueContactsAndHandles } from 'src/modules/auto-companies-and-contacts-creation/utils/get-unique-contacts-and-handles.util';
+import { Contact } from 'src/modules/contact-creation-manager/types/contact.type';
+import { getUniqueContactsAndHandles } from 'src/modules/contact-creation-manager/utils/get-unique-contacts-and-handles.util';
 
 describe('getUniqueContactsAndHandles', () => {
   it('should return empty arrays when contacts is empty', () => {
