+++ packages/twenty-server/src/modules/contact-creation-manager/utils/filter-out-contacts-from-company-or-workspace.util.ts
@@ -1,6 +1,6 @@
-import { Contact } from 'src/modules/auto-companies-and-contacts-creation/types/contact.type';
-import { getDomainNameFromHandle } from 'src/modules/auto-companies-and-contacts-creation/utils/get-domain-name-from-handle.util';
 import { ConnectedAccountWorkspaceEntity } from 'src/modules/connected-account/standard-objects/connected-account.workspace-entity';
+import { Contact } from 'src/modules/contact-creation-manager/types/contact.type';
+import { getDomainNameFromHandle } from 'src/modules/contact-creation-manager/utils/get-domain-name-from-handle.util';
 import { WorkspaceMemberWorkspaceEntity } from 'src/modules/workspace-member/standard-objects/workspace-member.workspace-entity';
 
 export function filterOutSelfAndContactsFromCompanyOrWorkspace(
