+++ packages/twenty-server/src/modules/contact-creation-manager/utils/get-company-name-from-domain-name.util.ts
