+++ packages/twenty-server/src/modules/contact-creation-manager/utils/get-domain-name-from-handle.util.ts
