+++ packages/twenty-server/src/modules/messaging/message-import-manager/drivers/gmail/services/messaging-save-messages-and-enqueue-contact-creation.service.ts
@@ -9,11 +9,11 @@ import { InjectMessageQueue } from 'src/engine/integrations/message-queue/decora
 import { MessageQueue } from 'src/engine/integrations/message-queue/message-queue.constants';
 import { MessageQueueService } from 'src/engine/integrations/message-queue/services/message-queue.service';
 import { WorkspaceDataSourceService } from 'src/engine/workspace-datasource/workspace-datasource.service';
+import { ConnectedAccountWorkspaceEntity } from 'src/modules/connected-account/standard-objects/connected-account.workspace-entity';
 import {
     CreateCompanyAndContactJob,
     CreateCompanyAndContactJobData,
-} from 'src/modules/auto-companies-and-contacts-creation/jobs/create-company-and-contact.job';
-import { ConnectedAccountWorkspaceEntity } from 'src/modules/connected-account/standard-objects/connected-account.workspace-entity';
+} from 'src/modules/contact-creation-manager/jobs/create-company-and-contact.job';
 import { MessagingMessageService } from 'src/modules/messaging/common/services/messaging-message.service';
 import {
     MessageChannelContactAutoCreationPolicy,
